move color docs
