+++ src/docs/color/background/background.tsx
@@ -1,5 +1,5 @@
-import { background, border, Table, text } from "../../core";
-import { ColorSample } from "../../old-docs/color/sample/sample";
+import { background, border, Table, text } from "../../../core";
+import { ColorSample } from "../sample/sample";
 import s from "./background.module.css";
 
 type BackgroundKey = keyof typeof background;
@@ -10,14 +10,14 @@ interface Row {
 
 const MakeColumn =
   (theme: "light" | "dark", text: string) =>
-  (row: Row): JSX.Element => (
-    <div className={theme}>
-      <ColorSample
-        background={background[row.key]}
-        foreground={{ type: "text", cls: text, usage: "both" }}
-      />
-    </div>
-  );
+    (row: Row): JSX.Element => (
+      <div className={theme}>
+        <ColorSample
+          background={background[row.key]}
+          foreground={{ type: "text", cls: text, usage: "both" }}
+        />
+      </div>
+    );
 
 const LightStrong = MakeColumn("light", text.normal);
 const LightWeak = MakeColumn("light", text.muted);
