+++ src/docs/color/category/category.mdx
