+++ src/docs/color/static/grid.tsx
