+++ src/docs/color/static/sample.module.css
