+++ src/docs/color/static/sample.tsx
@@ -1,6 +1,6 @@
 import Color from "color";
 import { useEffect, useRef, useState } from "react";
-import { text } from "../../../../core/src";
+import { text } from "../../../core";
 import s from "./sample.module.css";
 
 interface Props {
