+++ src/docs/color/static/static.mdx
