+++ src/docs/color/static/table.module.css
