+++ src/docs/color/static/table.tsx
@@ -1,4 +1,4 @@
-import { border, Table } from "../../../../core/src";
+import { border, Table } from "../../../core";
 import { ColorStaticSample } from "./sample";
 import s from "./table.module.css";
 
@@ -8,17 +8,17 @@ interface Props {
 
 const Color =
   (props: Props) =>
-  (row: number): JSX.Element => (
-    // eslint-disable-next-line react/prop-types
-    <ColorStaticSample name={`${props.name}-${row}`} />
-  );
+    (row: number): JSX.Element => (
+      // eslint-disable-next-line react/prop-types
+      <ColorStaticSample name={`${props.name}-${row}`} />
+    );
 
 const Name =
   (props: Props) =>
-  (row: number): JSX.Element => (
-    // eslint-disable-next-line react/prop-types
-    <span children={`${props.name}-${row}`} />
-  );
+    (row: number): JSX.Element => (
+      // eslint-disable-next-line react/prop-types
+      <span children={`${props.name}-${row}`} />
+    );
 
 export const ColorStaticTable = (props: Props): JSX.Element => (
   <div className={[s.container, border.weak].join(" ")}>
