+++ src/docs/color/text/text.mdx
