+++ src/docs/color/text/text.tsx
@@ -1,4 +1,4 @@
-import { background, border, Table, text } from "../../../../core/src";
+import { background, border, Table, text } from "../../../core";
 import { ColorSample, ColorSampleUsage } from "../sample/sample";
 import s from "./text.module.css";
 
@@ -19,18 +19,18 @@ interface Row {
 
 const MakeColumn =
   (theme: "light" | "dark", back: string) =>
-  (row: Row): JSX.Element => (
-    <div className={theme}>
-      <ColorSample
-        background={back}
-        foreground={{
-          type: "text",
-          cls: text[row.key],
-          usage: row.usage,
-        }}
-      />
-    </div>
-  );
+    (row: Row): JSX.Element => (
+      <div className={theme}>
+        <ColorSample
+          background={back}
+          foreground={{
+            type: "text",
+            cls: text[row.key],
+            usage: row.usage,
+          }}
+        />
+      </div>
+    );
 
 const LightStrong = MakeColumn("light", background.strong);
 const LightWeak = MakeColumn("light", background.weak);
